# swagger.yaml
openapi: 3.0.0
info:
  title: Jobs API
  contact: {}
  version: "1.0"
servers:
  - url: https://hope-jobs-api.herokuapp.com/api/v1
    variables: {}
paths:
...
  /jobs:
    post:
      tags:
        - Jobs
      summary: Create Jobs
      operationId: CreateJobs
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/CreateJobsRequest"
                - example:
                    company: Airtel
                    position: Customer Care Agent
            example:
              company: Airtel
              position: Customer Care Agent
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []
    get:
      tags:
        - Jobs
      summary: Get All Jobs
      operationId: GetAllJobs
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []
...
components:
  schemas:
...
    CreateJobsRequest:
      title: CreateJobsRequest
      required:
        - company
        - position
      type: object
      properties:
        company:
          type: string
        position:
          type: string
      example:
        company: Airtel
        position: Customer Care Agent
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...